requirements.txt
discord.py==1.7.2
pydrive2==1.8.2
psutil==5.8.0
beautifulsoup4==4.9.3
tweepy==3.10.0
pillow==8.2.0
leather==0.3.3
cairosvg==2.5.2